+++ src/@3rdweb-sdk/react/hooks/useRegistry.ts
@@ -11,8 +11,9 @@ import { useMemo } from "react";
 import { getAllMultichainRegistry } from "dashboard-extensions/common/read/getAllMultichainRegistry";
 import { polygon } from "thirdweb/chains";
 import { getContract } from "thirdweb";
-import { defineDashboardChain, thirdwebClient } from "lib/thirdweb-client";
+import { thirdwebClient } from "lib/thirdweb-client";
 import { BasicContract } from "contract-ui/types/types";
+import { defineDashboardChain } from "lib/v5-adapter";
 
 const MULTICHAIN_REGISTRY_ADDRESS =
   "0xcdAD8FA86e18538aC207872E8ff3536501431B73";
