+++ src/app/(dashboard)/(chain)/[chain_id]/layout.tsx
@@ -1,10 +1,8 @@
 /* eslint-disable react/forbid-dom-props */
 import { cn } from "@/lib/utils";
 import { ArrowLeftIcon, CircleAlertIcon } from "lucide-react";
-import { ToolTipLabel } from "@/components/ui/tooltip";
 import { Separator } from "@/components/ui/separator";
 import { ChainPageTabs } from "./tabs";
-import { ChainIcon } from "./ChainIcon";
 import { ChainOverview } from "./overview/OverviewPage";
 import Link from "next/link";
 import { StarButton } from "../components/client/star-button";
@@ -12,6 +10,7 @@ import { getChain } from "./utils";
 import { Metadata } from "next";
 import { getAbsoluteUrl } from "lib/vercel-utils";
 import { redirect } from "next/navigation";
+import { ChainIcon } from "./components/server/chain-icon";
 
 export async function generateMetadata(
   { params }: { params: { chain_id: string } },
@@ -59,45 +58,55 @@ export default async function ChainPageLayout({
   const isDeprecated = chain.status === "deprecated";
 
   return (
-    <section className="flex flex-col h-full gap-8">
-      {/* Header */}
-      <header
-        className="py-10 md:pt-10 md:pb-16 border-b relative bg-black overflow-hidden"
-        style={{
-          background: `hsl(var(--secondary))`,
-        }}
-      >
-        <div className="container px-4">
-          <Link
-            href="/chainlist"
-            className="inline-flex items-center gap-1 text-muted-foreground hover:text-foreground transition-colors"
-          >
-            <ArrowLeftIcon className="size-5" />
-            Chainlist
-          </Link>
-          <div className="h-4" />
-
-          <div className="flex gap-3 md:gap-5 items-center">
-            {chain.icon?.url && (
-              <ChainIcon
-                iconUrl={chain.icon.url}
-                className="hidden md:block size-[84px] bg-secondary p-2 border-2 rounded-full"
-              />
-            )}
-
-            {/* Chain Name */}
-
-            <h1
-              className={cn(
-                "font-semibold tracking-tighter text-4xl md:text-6xl",
-              )}
+    <>
+      {isDeprecated && (
+        <>
+          <div className="bg-destructive">
+            <div className="container px-4 py-4 flex flex-row items-center gap-4 text-white">
+              <CircleAlertIcon className="size-6 flex-shrink-0" />
+              <h3 className="font-semibold">
+                This chain has been marked as deprecated. Some or all services
+                may no longer function.
+              </h3>
+            </div>
+          </div>
+          <Separator />
+        </>
+      )}
+      <section className="flex flex-col h-full gap-8">
+        {/* Header */}
+        <header className="py-10 md:pt-10 md:pb-16 border-b relative overflow-hidden bg-secondary">
+          <div className="container px-4">
+            <Link
+              href="/chainlist"
+              className="inline-flex items-center gap-1 text-muted-foreground hover:text-foreground transition-colors"
             >
-              {chain.name}
-            </h1>
+              <ArrowLeftIcon className="size-5" />
+              Chainlist
+            </Link>
+            <div className="h-4" />
 
-            {/* Desktop tags */}
-            <div className="hidden md:flex text-base items-center gap-3">
-              {/* {isVerified && (
+            <div className="flex gap-3 md:gap-5 items-center">
+              {chain.icon?.url && (
+                <ChainIcon
+                  iconUrl={chain.icon.url}
+                  className="hidden md:block size-[84px] bg-secondary p-2 border-2 rounded-full"
+                />
+              )}
+
+              {/* Chain Name */}
+
+              <h1
+                className={cn(
+                  "font-semibold tracking-tighter text-4xl md:text-6xl",
+                )}
+              >
+                {chain.name}
+              </h1>
+
+              {/* Desktop tags */}
+              <div className="hidden md:flex text-base items-center gap-3">
+                {/* {isVerified && (
                   <ToolTipLabel label="Verified">
                     <Verified className="text-primary-foreground size-[36px]" />
                   </ToolTipLabel>
@@ -109,44 +118,28 @@ export default async function ChainPageLayout({
                   </ToolTipLabel>
                 )} */}
 
-              {isDeprecated && (
-                <ToolTipLabel label="Deprecated">
-                  <CircleAlertIcon className="text-destructive-foreground size-[36px]" />
-                </ToolTipLabel>
-              )}
+                <StarButton
+                  chainId={chain.chainId}
+                  iconClassName="size-[36px]"
+                />
+              </div>
 
-              <StarButton chainId={chain.chainId} iconClassName="size-[36px]" />
-            </div>
-
-            {/* Mobile star */}
-            <div className="md:hidden flex items-center">
-              <StarButton chainId={chain.chainId} iconClassName="size-6" />
-            </div>
-          </div>
-
-          {/* Mobile tags */}
-          {isDeprecated && (
-            <div className="md:hidden ">
-              <Separator className="my-5" />
-              <div className="flex flex-col gap-3">
-                {isDeprecated && (
-                  <div className="flex gap-3">
-                    <CircleAlertIcon className="text-destructive-foreground size-6" />
-                    <p> Deprecated </p>
-                  </div>
-                )}
+              {/* Mobile star */}
+              <div className="md:hidden flex items-center">
+                <StarButton chainId={chain.chainId} iconClassName="size-6" />
               </div>
             </div>
-          )}
-        </div>
-      </header>
 
-      <main className="container px-4 pb-20 flex-1">
-        <ChainOverview chain={chain} />
-        <div className="h-14" />
-        <ChainPageTabs chainSlug={params.chain_id} chain={chain} />
-        <div className="pt-8">{children}</div>
-      </main>
-    </section>
+            {/* Mobile tags */}
+          </div>
+        </header>
+        <main className="container px-4 pb-20 flex-1">
+          <ChainOverview chain={chain} />
+          <div className="h-14" />
+          <ChainPageTabs chainSlug={params.chain_id} chain={chain} />
+          <div className="pt-8">{children}</div>
+        </main>
+      </section>
+    </>
   );
 }
